+++ test/helper/dependency.js
@@ -1,20 +1,18 @@
 const assert = require('power-assert')
 const expect = require('expect')
-const { resolve, join } = require('path')
+const { resolve } = require('path')
 const dependency = require('../../lib/helper/dependency')
 
 describe('dependency', () => {
   it('error main', () => {
-    const entry = resolve(__dirname, '../fixtures/react')
-    const main = 'notexist.js'
+    const entry = resolve(__dirname, '../fixtures/react/notexist.js')
 
-    expect(() => dependency(entry, main))
-      .toThrow(`File path error: ${join(entry, main)}`)
+    expect(() => dependency(entry))
+      .toThrow(`File path error: ${entry}`)
   })
 
   it('react', () => {
-    const entry = resolve(__dirname, '../fixtures/react')
-    const main = 'index.js'
+    const entry = resolve(__dirname, '../fixtures/react/index.js')
     const trees = [
       'index.js',
       'components/index.js',
@@ -24,22 +22,20 @@ describe('dependency', () => {
       'dir.js',
     ]
 
-    assert(dependency(entry, main).map(({ path }) => path).join() === trees.join())
+    assert(dependency(entry).map(({ path }) => path).join() === trees.join())
   })
 
   it('vue', () => {
-    const entry = resolve(__dirname, '../fixtures/vue')
-    const main = 'index.js'
+    const entry = resolve(__dirname, '../fixtures/vue/index.js')
     const trees = ['index.js', 'main.vue']
 
-    assert(dependency(entry, main).map(({ path }) => path).join() === trees.join())
+    assert(dependency(entry).map(({ path }) => path).join() === trees.join())
   })
 
   it('ts', () => {
-    const entry = resolve(__dirname, '../fixtures/ts')
-    const main = 'index.tsx'
+    const entry = resolve(__dirname, '../fixtures/ts/index.tsx')
     const trees = ['index.tsx', 'hello.tsx']
 
-    assert(dependency(entry, main).map(({ path }) => path).join() === trees.join())
+    assert(dependency(entry).map(({ path }) => path).join() === trees.join())
   })
 })